+++ magpie/api/webhooks.py
@@ -4,7 +4,6 @@ from typing import TYPE_CHECKING
 
 import requests
 import transaction
-from pyramid.threadlocal import get_current_registry
 from six.moves.urllib.parse import urlparse
 
 from magpie import models
@@ -12,79 +11,119 @@ from magpie.api.schemas import UserStatuses
 from magpie.constants import get_constant
 from magpie.db import get_db_session_from_config_ini
 from magpie.register import get_all_configs
-from magpie.utils import ExtendedEnum, get_logger, get_settings, raise_log
+from magpie.utils import CONTENT_TYPE_JSON, FORMAT_TYPE_MAPPING, ExtendedEnum, get_logger, get_settings, raise_log
 
 if TYPE_CHECKING:
-    from typing import Dict, List, Optional
+    from typing import List, Optional
 
-    from magpie.typedefs import JSON, SettingsType, Str
+    from magpie.typedefs import (
+        AnySettingsContainer,
+        SettingsType,
+        Str,
+        WebhookConfig,
+        WebhookConfigSettings,
+        WebhookPayload,
+        WebhookTemplateParameters
+    )
 
 # List of keys that should be found for a single webhook item in the config
-WEBHOOK_KEYS = {
+WEBHOOK_KEYS_REQUIRED = {
     "name",
     "action",
     "method",
     "url",
     "payload"
 }
+WEBHOOK_KEYS_OPTIONAL = {
+    "format"
+}
+WEBHOOK_KEYS = WEBHOOK_KEYS_REQUIRED | WEBHOOK_KEYS_OPTIONAL
 
-# These are the parameters permitted to use the template form in the webhook payload.
-WEBHOOK_TEMPLATE_PARAMS = ["user_name", "tmp_url"]
+# These are *potential* parameters permitted to use the template form in the webhook payload.
+# Each parameter transferred to any given webhook are provided distinctively for each case.
+WEBHOOK_TEMPLATE_PARAMS = ["user_name", "user_id", "user_email", "callback_url"]
 
-HTTP_METHODS = ["GET", "HEAD", "POST", "PUT", "PATCH", "DELETE"]
+WEBHOOK_HTTP_METHODS = ["GET", "HEAD", "POST", "PUT", "PATCH", "DELETE"]
 
 LOGGER = get_logger(__name__)
 
 
 class WebhookAction(ExtendedEnum):
     """
-    Actions supported by webhooks.
+    Supported :term:`Webhook` actions.
     """
+
     CREATE_USER = "create_user"
+    """Triggered when a new user gets successfully created."""
+
     DELETE_USER = "delete_user"
+    """Triggered when an existing user gets successfully deleted."""
 
 
-def process_webhook_requests(action, params, update_user_status_on_error=False):
+def process_webhook_requests(action, params, update_user_status_on_error=False, settings=None):
+    # type: (WebhookAction, WebhookTemplateParameters, bool, Optional[AnySettingsContainer]) -> None
     """
     Checks the config for any webhooks that correspond to the input action, and prepares corresponding requests.
 
     :param action: tag identifying which webhooks to use in the config
-    :param params: dictionary containing the required parameters for the request, they will replace templates
-                    found in the payload
-    :param update_user_status_on_error: update the user status or not in case of a webhook error
+    :param params:
+        Dictionary containing the required parameters and associated values for the request following the event action.
+        Parameters will replace *templates* found in the ``payload`` definition of the webhook.
+    :param update_user_status_on_error: update the user status or not in case of a webhook error.
+    :param settings: application settings where webhooks configuration can be retrieved.
     """
     # Check for webhook requests
-    webhooks = get_settings(get_current_registry())["webhooks"][action.value]
-    if len(webhooks) > 0:
+    settings = get_settings(settings, app=True)
+    webhooks = settings.get("webhooks", {})  # type: WebhookConfig
+    if not webhooks:
+        return
+    action_webhooks = webhooks[action]
+    if len(action_webhooks) > 0:
         # Execute all webhook requests
-        pool = multiprocessing.Pool(processes=len(webhooks))
-        args = [(webhook, params, update_user_status_on_error) for webhook in webhooks]
+        pool = multiprocessing.Pool(processes=len(action_webhooks))
+        args = [(webhook, params, update_user_status_on_error) for webhook in action_webhooks]
         pool.starmap_async(send_webhook_request, args)
 
 
-def replace_template(params, payload):
+def replace_template(params, payload, force_str=False):
+    # type: (WebhookTemplateParameters, WebhookPayload, bool) -> WebhookPayload
     """
     Replace each template parameter from the payload by its corresponding value.
 
     :param params: the values of the template parameters
     :param payload: structure containing the data to be processed by the template replacement
+    :param force_str: enforce string conversion of applicable fields where non-string values are detected.
     :return: structure containing the data with the replaced template parameters
     """
     if isinstance(payload, dict):
-        return {replace_template(params, key): replace_template(params, value)
+        return {replace_template(params, key, force_str=True): replace_template(params, value)
                 for key, value in payload.items()}
     if isinstance(payload, list):
         return [replace_template(params, value) for value in payload]
-    if isinstance(payload, str):
-        for template_param in WEBHOOK_TEMPLATE_PARAMS:
-            if template_param in params:
-                payload = payload.replace("{" + template_param + "}", params[template_param])
+    if isinstance(payload, str):  # template fields are always string since '{<param>}' must be provided
+        for template_param in params:
+            template_replace = "{" + template_param + "}"
+            if template_param in WEBHOOK_TEMPLATE_PARAMS and template_replace in payload:
+                template_value = params[template_param]
+                # if result field is not a string and template is defined as is, allow value type replacement
+                if not force_str and not isinstance(template_value, str) and payload == template_replace:
+                    payload = template_value
+                # otherwise, enforce convert to string to avoid failing string replacement,
+                # but remove any additional quotes that might be defined to enforce non-string to string conversion
+                else:
+                    template_single_string = "'" + template_replace + "'"
+                    template_double_string = "\"" + template_replace + "\""
+                    for template_str in [template_single_string, template_double_string]:
+                        if payload == template_str and not isinstance(template_value, str):
+                            template_replace = template_str
+                    payload = payload.replace(template_replace, str(template_value))
         return payload
     # For any other type, no replacing to do
     return payload
 
 
 def send_webhook_request(webhook_config, params, update_user_status_on_error=False):
+    # type: (WebhookConfigSettings, WebhookTemplateParameters, bool) -> None
     """
     Sends a single webhook request using the input config.
 
@@ -95,14 +134,16 @@ def send_webhook_request(webhook_config, params, update_user_status_on_error=Fal
     """
     try:
         # Replace template parameters if a corresponding value was defined in input and send the webhook request
-        resp = requests.request(webhook_config["method"],
-                                replace_template(params, webhook_config["url"]),
-                                json=replace_template(params, webhook_config["payload"]))
+        ctype = FORMAT_TYPE_MAPPING.get(webhook_config.get("format", "json"), CONTENT_TYPE_JSON)
+        headers = {"Content-Type": ctype}
+        data = replace_template(params, webhook_config["payload"])
+        data_kw = {"json": data} if ctype == CONTENT_TYPE_JSON else {"data": data}  # json parsing error using 'data'
+        resp = requests.request(webhook_config["method"], webhook_config["url"], headers=headers, **data_kw)
         resp.raise_for_status()
     except Exception as exception:
         LOGGER.error("An exception has occurred with the webhook request : %s", webhook_config["name"])
         LOGGER.error(str(exception))
-        if "user_name" in params.keys() and update_user_status_on_error:
+        if "user_name" in params and update_user_status_on_error:
             webhook_update_error_status(params["user_name"])
 
 
@@ -124,14 +165,14 @@ def setup_webhooks(config_path, settings):
     """
 
     settings["webhooks"] = defaultdict(lambda: [])
-    webhooks_conf = settings["webhooks"]  # type: Dict[str, List[JSON]]
+    webhooks_conf = settings["webhooks"]  # type: WebhookConfig
     if not config_path:
         LOGGER.info("No configuration file provided to load webhook definitions.")
     else:
         LOGGER.info("Loading provided configuration file to setup webhook definitions.")
         webhook_configs = get_all_configs(config_path, "webhooks", allow_missing=True)
 
-        for cfg in webhook_configs:  # type: JSON
+        for cfg in webhook_configs:  # type: List[WebhookConfigSettings]
             for webhook in cfg:
                 # Validate the webhook config
                 if not isinstance(webhook, dict):
@@ -140,7 +181,7 @@ def setup_webhooks(config_path, settings):
                         exception=ValueError, logger=LOGGER
                     )
                 LOGGER.debug("Validating webhook: %s", webhook.get("name", "<undefined-name>"))
-                if set(webhook.keys()) != WEBHOOK_KEYS or not all(value for value in webhook.values()):
+                if set(webhook) != WEBHOOK_KEYS_REQUIRED or not all(value for value in webhook.values()):
                     raise_log(
                         "Missing or invalid key/value in webhook config from the config file {}".format(config_path),
                         exception=ValueError, logger=LOGGER
@@ -150,7 +191,7 @@ def setup_webhooks(config_path, settings):
                         "Invalid action {} found in webhook from config file {}".format(webhook["action"], config_path),
                         exception=ValueError, logger=LOGGER
                     )
-                if webhook["method"] not in HTTP_METHODS:
+                if webhook["method"] not in WEBHOOK_HTTP_METHODS:
                     raise_log(
                         "Invalid method {} found in webhook from config file {}".format(webhook["method"], config_path),
                         exception=ValueError, logger=LOGGER
@@ -163,5 +204,6 @@ def setup_webhooks(config_path, settings):
                     )
 
                 # Regroup webhooks by action key
-                webhook_sub_config = {k: webhook[k] for k in set(list(webhook.keys())) - {"action"}}
-                webhooks_conf[webhook["action"]].append(webhook_sub_config)
+                webhook_cfg = {k: webhook[k] for k in WEBHOOK_KEYS if k in webhook}  # noqa # ignore optional fields
+                webhook_action = WebhookAction.get(webhook["action"])
+                webhooks_conf[webhook_action].append(webhook_cfg)